apps/producer/requirements.txt
aiokafka>=0.10
fastavro>=1.9
httpx>=0.27
pydantic>=2.9
python-dotenv>=1.0
uvloop; platform_system != "Windows"
